requirements.txt
groq
python-dotenv
'httpx[socks]'